Fix: implement create_sim_directory_map for ContainerPlatform
